fix: moved get_doctype_fields logic to crm/api/doc.py
