+++ crm/fcrm/doctype/crm_lead/crm_lead.json
@@ -13,7 +13,6 @@
   "website",
   "industry",
   "job_title",
-  "naming_series",
   "source",
   "person_tab",
   "salutation",
@@ -23,6 +22,7 @@
   "mobile_no",
   "organization_tab",
   "section_break_uixv",
+  "naming_series",
   "lead_name",
   "middle_name",
   "gender",
@@ -199,7 +199,7 @@
  "image_field": "image",
  "index_web_pages_for_search": 1,
  "links": [],
- "modified": "2023-11-21 13:02:11.680600",
+ "modified": "2023-11-22 13:03:02.261001",
  "modified_by": "Administrator",
  "module": "FCRM",
  "name": "CRM Lead",
